관리자 상품 등록 구현 중 (이미지 압축, 변환 구현)

+++ src/main/java/command/manager/ManagerUserDeleteProcessAction.java
@@ -14,21 +14,23 @@ public class ManagerUserDeleteProcessAction implements CommandAction {
 
 	@Override
 	public String requestProcess(HttpServletRequest request, HttpServletResponse response) throws Throwable {
-		request.setCharacterEncoding("utf-8");
+		if ((boolean) request.getSession().getAttribute("admin")) {
+			request.setCharacterEncoding("utf-8");
 
-		StringBuilder sb = new StringBuilder();
-		try (BufferedReader reader = request.getReader()) {
-			String line = null;
-			while ((line = reader.readLine()) != null) {
-				sb.append(line);
+			StringBuilder sb = new StringBuilder();
+			try (BufferedReader reader = request.getReader()) {
+				String line = null;
+				while ((line = reader.readLine()) != null) {
+					sb.append(line);
+				}
 			}
+
+			JSONObject jsonRequest = new JSONObject(sb.toString());
+			int member_id = jsonRequest.getInt("member_id");
+
+			MemberDAO mdao = MemberDAO.getInstance();
+			mdao.deleteMember(member_id);
 		}
-
-		JSONObject jsonRequest = new JSONObject(sb.toString());
-		int member_id = jsonRequest.getInt("member_id");
-
-		MemberDAO mdao = MemberDAO.getInstance();
-		mdao.deleteMember(member_id);
 		return null;
 	}
 }
